# Conflicts:
#	app/src/main/java/com/project/passwordmanager/adapters/CredentialsAdapter.kt
#	app/src/main/java/com/project/passwordmanager/fragments/AddCredentialDialogFragment.kt
#	app/src/main/java/com/project/passwordmanager/fragments/UnlockDialogFragment.kt

+++ app/src/main/java/com/project/passwordmanager/adapters/WidgetConfigurationCredentialsAdapter.kt
@@ -22,7 +22,7 @@ class WidgetConfigurationCredentialsAdapter(context: Context):
             notifyDataSetChanged()
         }
 
-    val selectedCredentialsIds: MutableList<Long> = mutableListOf()
+    val selectedCredentialsIds: MutableList<Int> = mutableListOf()
 
     inner class WidgetConfigurationCredentialsViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
     {
